Hadupils::Assets to abstract files, jars, tarballs
Basic classes to represent the main kinds of assets we need to
use within the hadoop distributed cache:
* Hadupils::Assets::Archive for tarballs w/ ".tar.gz" extension
* Hadupils::Assets::Jar for jars (".jar")
* Hadupils::Assets::File for all other regular files

Each knows its hive command for adding to a hive session.

Each distinguishes between name as it would appear in the hadoop
distributed cache (the file basename) and its full path.

The Hadupils::Assets.asset_for function gives an asset instance
for a specified file.