Hadupils::Runners module for running hadoop apps
Hadupils::Runners::Hive knows how to run the underlying
hive command.  It assumes by default that hive lives at
$HIVE_HOME/bin/hive, and understands how to prepare options
for that command.

Parameters can be given to the runner as a list of things that
either #to_s nicely (and correspond directly to hive options)
or as things that have #hive_opts (a list of hive option strings).

TO-DO: add a Hadupils::Runners::Hadoop for driving hadoop calls.