Hadupils::Extensions::Static for basic config dirs
A Hadupils::Extensions::Static instance wraps a directory of
static configuration files.  It does not gather assets.  It does
not calculate hiverc contents.  It expects to find static config
or init files within the directory.  It provides (for now) a
hivercs method to wrap a static "hiverc" file if one exists.