Broader unit test for the hive command
With minimal stubbing and custom assembly of a temporary directory
including both a "user config" and a "hadoop-ext", new unit test
in test/unit/commands_test.rb verifies that the
Hadupils::Commands::Hive class assembles the correct system call.
It verifies that the hivercs provided to the hive call have the
expected content.

The test is done with a Mocha expectation block and as such is not
very informative upon failure.  But it does exercise a fairly broad
swath of functionality within the library.