Hadupils::Extensions::HiveSet implementation
The HiveSet allows a directory to act as a collection of
Hadupils::Extensions::Hive subdirectories, and provides a
convenient interface for aggregating all jars and initialization
files and whatnot across the collection.

Also added a hive_aux_jars_path method to the Hive (and HiveSet)
classes to make it easy to use in preparing calls to hive.