Fixed odd hanging bug in build_archive function
The switch to Open3.popen3 for driving the tar call to bundle
up the hive-ext archive was hanging when reading from stderr,
which is odd.

Switched to popen2 which ignores stderr and now the hanging ain't
hangin'.

I should have tried to reproduce this in a test, but I should have
done a lot of things by my age.